hardcode @jest/globals API + add test
